Migrated gender to sex with use of enum
All references of a string gender will now refer to the enum value
with male, female and other. The factory also correctly generates
a random sex with the help of faker.

User profile update now does a direct null or undefined check
instead of a boolean check. As 0 based boolean values like enums
will be ignored.

 * Removed no longer required comment about gender
 * Changed 'userProfileTemplate' to 'userProfileSettings'