feat(authentication): authentication now uses JWT
Switching the use of random strings for properly implemented json web tokens, these tokens are now verified and then verified on the account before being classes as authorized.

Once authorized, the user will be pushed on to the request object. Any request that has the middleware to be authorized can be confident that the user object exists if it makes it to that request.

 * removed redunant await on getCustomRepository

 * secret environment used for token generation (this should be purely random string generted in production).