feat(user): toJSON
toJSON now exits for the user model, ensures that the token and the password of the object are fully removed before being sent to the client.

I decided to keep the sanitize method for more explicit sanitization processes but does now create a new user object and then remove the properties.

 * Authentication middleware calls toJSON to ensure there is not a long-running in-memory user with the token and password.