maint(game-routes): binding schedule and game through middleware
The game's and schedules are now being bound to the request object through
the middleware, allowing for the simplification of the requests. If the
:game or :schedule is specified and the binding middleware is used. Then
the IGame or ISchedule request interfaces can be specified for the request
to gain access.

If the request make it through the middleware, you can ensure that the
object exists.

 * Emails are now sent on applications being applied and the footer is
 updated with the discord link.