maint(reset-email): reset email schema validation
Ensuring that the given email and password for resetting a users email
is valid. Ensuring that test coverage exists. This lead to fixing a bug
that was specifying ownership endpoint but due to the implementation the
change would only affect the current user. Thus it only caused a missing
id error.

 * Improved authentication failed error messages.