feature: binding user object by id via a middleware
non-authenticated users can be bound to the request object via the
bindUserFromUserParam onto the request object. This ensures that the
given user exists and the given id is a actual number.

This can be later accessed via the IUserRequest object after binding
from the middleware. Set for request.boundUser instead of request.user
which is currently reserved for authenticated users.

 * Added api docs for the user.controller.