fix: resigning and oauth limits
When a user resigns, the application now correctly deletes its related
information. It was caused my not using the correct database connection
and the change not taking place.

Oauth api is now moderation only with specified limits and offsets to
gather a range of current linked accounts for all users.

 * A endpoint needs to be created to allow gather the given users linked
accounts directly.