fix: activated game schedules links are now correctly bound
When a game schedule is activated, the game is created and the link
between the game and the game schedules are now fixed. This means that
there is a one to one relationship between both the game and the
schedule. This allows gathering the users related games to correctly
perform a deletion.

 * Added test templates and implemented authentication/missing schedule
test cases.