feature: include players with game request /games/id?players=true
You can now specify if you want to include complete user information for
the players when gathering the game. By default no player information
will be gathered. Only the already existing id, team and username will
be returned.

 * Boolean parser now included within the helper, used to process query
booleans with a deafult value if parse could not be determined.
 * test cases for gathering with and without the players complete
details.
 *  bindGameFromParam uses helper method over raw checks.