fix: profile updates and sex enum
When updating profiles, null values are now allowed which is the default
value within the database. Also the sex is now clearly defined as
integers within the enum and the schema. Ensuring not to allow string
based sex as a input for profile update.

If strings are allowed, when they insert into the database, it causes a
query error that 500's.