feat: shifting coins to user statistics
When a user links there Twitch account with DevWars, if they had a existing twitch account with coins on it but not linked, all coins will be transferred onto the users statistics (coins).  The coins value will then be removed from the twitch storage.

All following updates will be done to the user stats. If the user decides to disconnect there twitch account, all the coins will then be continued to be added to the linked account object (as before) until linked again.

 * Ensured to use the repository with this instead of using the static object.
 * npm run audit.
 * updated tests to reflect changes.