maint: coins are no longer shifted on account link
When a linked account is removed, if storage has content, it will not be deleted just the removal of the relation. But otherwise deleted, since a new linked account would be no different.

User stats endpoint will now gather all the coins for all linked accounts if they exist. Returning a up to date coin count, this is the same for /coins endpoint.

 * Logger will now correctly print objects.
 * performing a commit will now also perform a lint.