refactor: no longer require removing editor ids
When deleting a user, we keep the ids the same for the editor and index on the player object. Only setting the internal Id to null. This allow the front end to be able to handle multiple deleted users while also showing which user had which editor (removed or not).