feat: game storage interfacing
Provided game storage based interfacing to ensure that when trying to
work with the game storage on the game object dirctly that the user is
aware of the types and what is expected to exist and not exist.

This does increase management by ensuring to maintain the interface but
massively improves development time.